pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
authors = ["Mikhail Egorov <pe.egorov@gmail.com>"]
readme = "README.md"
classifiers = [
  "Environment :: Console",
  "Intended Audience :: Other Audience",
  "License :: Free For Home Use",
  "Programming Language :: Python :: 3.10",
  "Topic :: Text Processing"
]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.main:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.4"
coverage = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
